Refactored all classes and added WH Appium Tests
