+++ src/test/java/HD_Pages/AllMarketQR.java
@@ -1,4 +1,4 @@
-package pages;
+package HD_Pages;
 
 import io.appium.java_client.android.AndroidDriver;
 import org.openqa.selenium.By;
