+++ src/test/java/HD_Pages/LoadTruck.java
@@ -1,4 +1,4 @@
-package pages;
+package HD_Pages;
 
 import io.appium.java_client.android.AndroidDriver;
 import org.openqa.selenium.By;
@@ -6,21 +6,24 @@ import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
 import org.testng.Assert;
 
-public class BoxSearch {
-
+public class LoadTruck {
+    /*
+    @author : Rahul Dash, Nikita Gopathi
+     */
     public void ClickNBack(WebDriverWait wait, AndroidDriver driver ,String packageName) {
         wait.until(ExpectedConditions.visibilityOfElementLocated(By.id(
-                packageName + ":id/home_box_search_img"))).click();
-        String boxSearchLabel = wait.until(ExpectedConditions.visibilityOfElementLocated(By.xpath(
-                "/hierarchy/android.widget.FrameLayout/android.widget.LinearLayout/android.widget.FrameLayout/android.widget.LinearLayout/android.widget.FrameLayout/android.widget.RelativeLayout/android.widget.RelativeLayout[1]/android.widget.TextView"))).getText();
-        System.out.println("Box Search header : "+boxSearchLabel);
-        Assert.assertTrue(boxSearchLabel.equals("Box Search"));
+                packageName + ":id/home_load_truck_img"))).click();
+        String loadTruckLabel = wait.until(ExpectedConditions.visibilityOfElementLocated(By.xpath(
+                "/hierarchy/android.widget.FrameLayout/android.widget.LinearLayout/android.widget.FrameLayout/android.widget.LinearLayout/android.widget.FrameLayout/android.widget.LinearLayout/android.widget.LinearLayout[1]/android.view.ViewGroup/android.widget.RelativeLayout/android.widget.TextView"))).getText();
+        System.out.println("Load Truck header : " + loadTruckLabel);
+        Assert.assertTrue(loadTruckLabel.equals("Load Truck"));
         wait.until(ExpectedConditions.visibilityOfElementLocated(By.id(
                 packageName + ":id/img_back"))).click();
-        //TODO - PUT IN WHAT YOU ARE TRYING TO VALIDATE HERE
         String hdLabel = wait.until(ExpectedConditions.visibilityOfElementLocated(By.id(
                 packageName + ":id/home_delivery_persona_label"))).getText();
         //TODO - actually get this test to work
         Assert.assertTrue(hdLabel.equals("Home Delivery"));
+
     }
+
 }
