+++ src/test/java/HD_Pages/ReportDamage.java
@@ -1,5 +1,4 @@
-
-package pages;
+package HD_Pages;
 
 import io.appium.java_client.android.AndroidDriver;
 import org.openqa.selenium.By;
@@ -7,18 +6,16 @@ import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
 import org.testng.Assert;
 
-public class ReportMissing {
-    /*@author -
-    Rahul Dash*/
-    public void ClickNBack(WebDriverWait wait, AndroidDriver driver, String packageName) {
+public class ReportDamage {
+    public void ClickNBack(WebDriverWait wait, AndroidDriver driver,String packageName) {
         wait.until(ExpectedConditions.visibilityOfElementLocated(By.id(
                 packageName + ":id/report_damage_img"))).click();
         wait.until(ExpectedConditions.visibilityOfElementLocated(By.id(
-                packageName + ":id/report_missing"))).click();
-        String reportMissingLabel = wait.until(ExpectedConditions.visibilityOfElementLocated(By.xpath(
+                packageName + ":id/damage_selection"))).click();
+        String reportDmgLabel = wait.until(ExpectedConditions.visibilityOfElementLocated(By.xpath(
                 "/hierarchy/android.widget.FrameLayout/android.widget.LinearLayout/android.widget.FrameLayout/android.widget.LinearLayout/android.widget.FrameLayout/androidx.drawerlayout.widget.DrawerLayout/android.view.ViewGroup/android.view.ViewGroup/android.widget.RelativeLayout/android.widget.RelativeLayout/android.widget.TextView"))).getText();
-        System.out.println("Report Missing header : "+reportMissingLabel);
-        Assert.assertTrue(reportMissingLabel.equals("Report Missing"));
+        System.out.println("Report Damage header : "+reportDmgLabel);
+        Assert.assertTrue(reportDmgLabel.equals("Report Damage"));
         wait.until(ExpectedConditions.visibilityOfElementLocated(By.id(
                 packageName + ":id/action_home"))).click();
         String hdLabel = wait.until(ExpectedConditions.visibilityOfElementLocated(By.id(
