+++ src/test/java/HD_Pages/Returns.java
@@ -1,13 +1,11 @@
-package pages;
+package HD_Pages;
 
-import io.appium.java_client.MobileElement;
 import io.appium.java_client.android.AndroidDriver;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
 import org.testng.Assert;
-
 import java.util.concurrent.TimeUnit;
 
 public class Returns {
