+++ src/test/java/KH_Pages/AllMarketReturnsQRCodes_KH.java
@@ -5,7 +5,6 @@ import org.openqa.selenium.By;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
 import org.testng.Assert;
-
 import java.util.concurrent.TimeUnit;
 
 public class AllMarketReturnsQRCodes_KH {
