+++ src/test/java/KH_Pages/Feedback_KH.java
@@ -1,0 +1,29 @@
+package KH_Pages;
+
+import io.appium.java_client.android.AndroidDriver;
+import org.openqa.selenium.By;
+import org.openqa.selenium.support.ui.ExpectedConditions;
+import org.openqa.selenium.support.ui.WebDriverWait;
+import org.testng.Assert;
+import java.util.concurrent.TimeUnit;
+
+public class Feedback_KH {
+    /*
+    @author : Nikita Gopathi
+     */
+    public void ClickNBack(WebDriverWait wait, AndroidDriver driver , String packageName) {
+        wait.until(ExpectedConditions.visibilityOfElementLocated(By.xpath(
+                "//android.widget.ImageButton[@content-desc=\"Open navigation drawer\"]"))).click();
+        driver.manage().timeouts().implicitlyWait(10, TimeUnit.SECONDS);
+        wait.until(ExpectedConditions.visibilityOfElementLocated(By.xpath(
+                "/hierarchy/android.widget.FrameLayout/android.widget.LinearLayout/android.widget.FrameLayout/android.widget.LinearLayout/android.widget.FrameLayout/androidx.drawerlayout.widget.DrawerLayout/android.widget.FrameLayout/androidx.recyclerview.widget.RecyclerView/androidx.appcompat.widget.LinearLayoutCompat[10]/android.widget.CheckedTextView"))).click();
+        String feedbackHeader= driver.findElementById(packageName+":id/feedback_header_txt").getText();
+        System.out.println("Feedback tab header: "+feedbackHeader);
+        Assert.assertTrue(feedbackHeader.equals("Feedback"));
+        wait.until(ExpectedConditions.visibilityOfElementLocated(By.id(
+                packageName+":id/action_home"))).click();
+        String khLabel = wait.until(ExpectedConditions.visibilityOfElementLocated(By.id(
+                packageName+":id/keyholder_persona_label"))).getText();
+        Assert.assertTrue(khLabel.equals("Keyholder"));
+    }
+}
