+++ src/test/java/KH_Pages/Returns_KH.java
@@ -6,7 +6,6 @@ import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
 import org.testng.Assert;
-
 import java.util.concurrent.TimeUnit;
 
 public class Returns_KH {
