+++ src/test/java/WH_Pages/FeedbackWH.java
@@ -5,10 +5,9 @@ import org.openqa.selenium.By;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
 import org.testng.Assert;
-
 import java.util.concurrent.TimeUnit;
 
-public class FeedBackWH {
+public class FeedbackWH {
     /*@author -
     Rahul Dash*/
     public void ClickNBack(WebDriverWait wait, AndroidDriver driver , String packageName) {
@@ -16,7 +15,7 @@ public class FeedBackWH {
                 "//android.widget.ImageButton[@content-desc=\"Open navigation drawer\"]"))).click();
         driver.manage().timeouts().implicitlyWait(10, TimeUnit.SECONDS);
         wait.until(ExpectedConditions.visibilityOfElementLocated(By.xpath(
-                "/hierarchy/android.widget.FrameLayout/android.widget.LinearLayout/android.widget.FrameLayout/android.widget.LinearLayout/android.widget.FrameLayout/androidx.drawerlayout.widget.DrawerLayout/android.view.ViewGroup/android.view.ViewGroup/android.widget.LinearLayout/android.widget.ScrollView/android.widget.LinearLayout/android.widget.LinearLayout[4]/android.widget.RelativeLayout[1]/android.widget.FrameLayout/android.widget.RelativeLayout/android.widget.ImageView"))).click();
+                "/hierarchy/android.widget.FrameLayout/android.widget.LinearLayout/android.widget.FrameLayout/android.widget.LinearLayout/android.widget.FrameLayout/androidx.drawerlayout.widget.DrawerLayout/android.widget.FrameLayout/androidx.recyclerview.widget.RecyclerView/androidx.appcompat.widget.LinearLayoutCompat[11]/android.widget.CheckedTextView"))).click();
         String FeedbackHead= driver.findElementById(packageName+":id/feedback_header_txt").getText();
         System.out.println("Feedback tab header: "+FeedbackHead);
         Assert.assertTrue(FeedbackHead.equals("Feedback"));
