+++ src/test/java/WH_Pages/InboundReceiptException_WH.java
@@ -6,10 +6,9 @@ import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
 import org.testng.Assert;
-
 import java.util.concurrent.TimeUnit;
 
-public class InboundReceiptException {
+public class InboundReceiptException_WH {
     /*
         @author : Rahul Dash
          */
