+++ src/test/java/WH_Pages/InventoryAudit_WH.java
@@ -5,10 +5,9 @@ import org.openqa.selenium.By;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
 import org.testng.Assert;
-
 import java.util.concurrent.TimeUnit;
 
-public class InventoryAudit {
+public class InventoryAudit_WH {
     /*
         @author : Rahul Dash
          */
