+++ src/test/java/WH_Pages/ReturnToADC_WH.java
@@ -5,10 +5,9 @@ import org.openqa.selenium.By;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
 import org.testng.Assert;
-
 import java.util.concurrent.TimeUnit;
 
-public class ReturnToADC {
+public class ReturnToADC_WH {
     /*
         @author : Rahul Dash
          */
